NN kinda working
Played around with having several output nodes, it works. So implemented that into my program.
There seems to be an overflow problem with my sigmoid function which is rather quite annoying since otherwise it the program seems to be working.
It only throws the error when I am trying to train it with more than 1 number. e.g. 0 and 1